# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InnoDD API",
    "version": "2.0.0",
    "description": "Complete InnoDD drug-discovery platform surface, generated from source (not the Postman collection).\n\n**Two surfaces:**\n- `/api/v1/*` — legacy internal pipeline modules (TxKG, Literature Mining, Drug Curation, Screening, Novelty Search) plus Auth/Users/Items scaffolding and job Orchestration. No X-Drp-Token needed.\n- `/v1/*` — the DRP frontend-facing contract (async job pattern: POST creates a job -> 202 {jobId}, then poll status/result). Needs `X-Drp-Token` (or `Authorization: Bearer`) from `/v1/auth/login`.\n\n**Known gaps / stubs (do not rely on for testing):**\n- `POST /api/v1/auth/login` and `/api/v1/auth/token` are unimplemented stubs (`pass`) in the legacy scaffold.\n- `/api/v1/users/*` imports a `UserRead` model that does not exist in `app/schemas/user.py` — this router likely fails to import and is NOT actually mounted.\n- `/api/v1/agentic/*` never mounts — `agent_routes.py` does not exist as a file; `_include()` swallows the ImportError.\n- Several module routers are double-mounted: Literature Mining, Drug Curation, Screening, and Novelty Search each exist BOTH under their versioned prefix (e.g. `/api/v1/literature-mining/...`) AND again with no prefix as a legacy alias (e.g. `/api/v1/search-keywords/`) — both paths are live simultaneously and hit the same handler.",
    "contact": { "name": "InnoDD API" }
  },
  "servers": [
...
  ],
  "components": {
    "securitySchemes": {
      "drpToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Drp-Token",
        "description": "Access token returned by POST /v1/auth/login or /v1/auth/refresh. The server also accepts a standard Authorization: Bearer header as a fallback."
      }
    },
    "schemas": {
...
      "NovSearchRequest": {
        "type": "object",
        "properties": {
          "source": { "type": "string", "nullable": true }, "query": { "type": "string", "nullable": true },
          "target": { "type": "string", "nullable": true }, "drug": { "type": "string", "nullable": true },
          "disease": { "type": "string", "nullable": true }, "candidate_id": { "type": "string", "nullable": true },
          "docking_id": { "type": "string", "nullable": true }, "num_results": { "type": "integer", "minimum": 1, "maximum": 20, "default": 5 }
        }
      },
      "NovSearchQARequest": {
        "type": "object",
        "required": ["question"],
        "properties": {
          "question": { "type": "string" },
          "patent_ids": { "type": "array", "items": { "type": "string" } },
          "top_k": { "type": "integer", "minimum": 1, "maximum": 50 }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/txkg/tools/resolve-disease": {
      "get": {
        "tags": ["1. TxKG — Discovery"], "summary": "§1 Resolve disease",
        "parameters": [{ "name": "disease", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/txkg/tools/paths": {
      "get": {
        "tags": ["1. TxKG — Discovery"], "summary": "§6 Path reconstruction (BFS)",
        "parameters": [
          { "name": "disease", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "uniprot_id", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "max_hops", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 5 } },
          { "name": "max_paths", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 25 } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/txkg/tools/score": {
      "get": {
        "tags": ["1. TxKG — Discovery"], "summary": "§2-§5 RWRH propagation + degree correction",
        "parameters": [
          { "name": "disease", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "top_known", "in": "query", "schema": { "type": "integer", "default": 25, "minimum": 1, "maximum": 200 } },
          { "name": "top_hidden", "in": "query", "schema": { "type": "integer", "default": 25, "minimum": 1, "maximum": 200 } },
          { "name": "hidden_threshold", "in": "query", "schema": { "type": "number" } },
          { "name": "keep_unconfirmed", "in": "query", "schema": { "type": "boolean", "default": true } },
          { "name": "reconstruct_paths", "in": "query", "schema": { "type": "boolean", "default": true } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/txkg/tools/evidence": {
      "get": {
        "tags": ["1. TxKG — Discovery"], "summary": "§7-§8 Novelty + interpretation evidence",
        "parameters": [
          { "name": "disease", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "uniprot_ids", "in": "query", "required": true, "schema": { "type": "string" }, "description": "CSV" },
          { "name": "articles_per_target", "in": "query", "schema": { "type": "integer", "default": 2, "minimum": 1, "maximum": 10 } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/txkg/discover": {
      "get": {
        "tags": ["1. TxKG — Discovery"], "summary": "Full §1-§10 discovery chain",
        "parameters": [
          { "name": "disease", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "top_known", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "top_hidden", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "interpret_top", "in": "query", "schema": { "type": "integer", "default": 5 } },
          { "name": "articles_per_target", "in": "query", "schema": { "type": "integer", "default": 2 } },
          { "name": "hidden_threshold", "in": "query", "schema": { "type": "number" } },
          { "name": "keep_unconfirmed", "in": "query", "schema": { "type": "boolean", "default": true } },
          { "name": "session_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/txkg/explain": {
      "get": {
        "tags": ["1. TxKG — Discovery"], "summary": "§11 Scoped follow-up",
        "parameters": [
          { "name": "session_id", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "uniprot_id", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/txkg/ask": {
      "get": {
        "tags": ["1. TxKG — Discovery"], "summary": "§11 Supervisor NL router",
        "parameters": [
          { "name": "message", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "session_id", "in": "query", "schema": { "type": "string", "default": "default" } },
          { "name": "top_known", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "top_hidden", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/graphs/discovery/{filename}": {
      "get": {
        "tags": ["1. TxKG — Discovery"], "summary": "Serve discovery graph HTML",
        "parameters": [{ "name": "filename", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^txkg_discovery_.*$" } }],
        "responses": { "200": { "description": "OK", "content": { "text/html": {} } } }
      }
    },
...
    "/api/v1/novsearch/run": {
      "post": {
        "tags": ["5. Novelty Search Agent (spec)"], "summary": "NovSearch run (legacy alias)",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NovSearchRequest" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/v1/novsearch/ask": {
      "post": {
        "tags": ["5. Novelty Search Agent (spec)"], "summary": "NovSearch QA (legacy alias)",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NovSearchQARequest" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}